feat: make peers page first render fast
The peers page was taking >2s to load on my machine. Turns out the
d3 generated svg of the world map was around 3MiB

This PR makes extracts the world map to a static svg that is used as
a background image. It's ~79KiB / ~20KiB gzippped and cached as a
static resource rather than generated by js.

The peers page now loads faster than all the others. There is still
too much fan spinning caused by the geo look ups, but this is an
improvement.

License: MIT
