+++ src/peers/WorldMap/WorldMap.js
@@ -4,23 +4,16 @@ import { connect } from 'redux-bundler-react'
 import { AutoSizer } from 'react-virtualized'
 import { translate } from 'react-i18next'
 import * as d3 from 'd3'
-import * as topojson from 'topojson'
-
-import worldData from './world.json'
+import staticMapSrc from './StaticMap.svg'
 
 const WorldMap = ({ t }) => {
   return (
-    <div className='flex w-100 mb4' style={{ 'height': '550px' }}>
-      <AutoSizer>
-        { ({ height, width }) => (
-          <GeoPath width={width} height={height}>
+    <div className='flex w-100 mb4' style={{ 'height': '550px', background: `transparent url(${staticMapSrc}) center` }}>
+      <AutoSizer disableHeight>
+        { ({ width }) => (
+          <GeoPath width={width} height={550}>
             { ({ path }) => (
-              <div className='relative'>
-                <Map height={height} width={width} path={path} />
-                <div className='absolute top-0'>
-                  <MapPins height={height} width={width} path={path} />
-                </div>
-              </div>
+              <MapPins height={550} width={width} path={path} />
             )}
           </GeoPath>
         )}
@@ -45,54 +38,6 @@ const GeoPath = ({ width, height, children }) => {
   const path = d3.geoPath().projection(projection)
 
   return children({ path })
-}
-
-// Earth! It's complicated, so try not to re-render so much
-const Map = ({ width, height, path }) => {
-  // https://github.com/d3/d3-geo/blob/master/README.md#geoGraticule
-  const graticule = d3.geoGraticule()
-
-  const el = d3.select(ReactFauxDOM.createElement('svg'))
-    .attr('width', width)
-    .attr('height', height)
-
-  // Fill Pattern
-  el.append('defs')
-    .append('pattern')
-    .attr('id', 'gridpattern')
-    .attr('x', 0)
-    .attr('y', 0)
-    .attr('width', 5)
-    .attr('height', 5)
-    .attr('patternUnits', 'userSpaceOnUse')
-    .append('circle')
-    .attr('cx', 3)
-    .attr('cy', 3)
-    .attr('r', 1)
-    .attr('fill', '#AAA')
-    .attr('stroke', '#DDD')
-
-  el.append('path')
-    .datum(graticule)
-    .attr('class', 'graticule')
-    .attr('d', path)
-
-  el.insert('path', '.graticule')
-    .datum(topojson.feature(worldData, worldData.objects.land))
-    .attr('d', path)
-    .attr('fill', 'url(#gridpattern)')
-
-  el.insert('path', '.graticule')
-    .datum(topojson.mesh(
-      worldData,
-      worldData.objects.countries,
-      (a, b) => a !== b
-    ))
-    .attr('d', path)
-    .attr('fill', 'none')
-    .attr('stroke', 'none')
-
-  return el.node().toReact()
 }
 
 // Just the dots on the map, this gets called a lot.
